# Speed up equalize transform: use bincount instead of histc