--- a/test/test_functional_tensor.py
+++ b/test/test_functional_tensor.py
@@ -984,10 +984,11 @@ class CUDATester(Tester):
         # TODO: when # https://github.com/pytorch/pytorch/issues/53194 is fixed,
         # only use bincount and remove that test.
         size = 1_000
-        img_chan = torch.randint(0, 256, size=(1_000,)).to('cpu')
+        img_chan = torch.randint(0, 256, size=size).to('cpu')
         scaled_cpu = F_t._scale_channel(img_chan)
         scaled_cuda = F_t._scale_channel(img_chan.to('cuda'))
-        self.assertTrue(scaled_cpu.equal(scaled_cuda))
+        self.assertTrue(scaled_cpu.equal(scaled_cuda.to('cpu')))
+
 
 if __name__ == '__main__':
     unittest.main()