--- a/test/test_functional_tensor.py
+++ b/test/test_functional_tensor.py
@@ -983,7 +983,7 @@ class CUDATester(Tester):
         """
         # TODO: when # https://github.com/pytorch/pytorch/issues/53194 is fixed,
         # only use bincount and remove that test.
-        size = 1_000
+        size = (1_000,)
         img_chan = torch.randint(0, 256, size=size).to('cpu')
         scaled_cpu = F_t._scale_channel(img_chan)
         scaled_cuda = F_t._scale_channel(img_chan.to('cuda'))