--- a/torchvision/transforms/functional_tensor.py
+++ b/torchvision/transforms/functional_tensor.py
@@ -886,7 +886,7 @@ def autocontrast(img: Tensor) -> Tensor:
 
 
 def _scale_channel(img_chan):
-    hist = torch.histc(img_chan.to(torch.float32), bins=256, min=0, max=255)
+    hist = torch.bincount(img_chan.view(-1),  minlength=256)
 
     nonzero_hist = hist[hist != 0]
     step = nonzero_hist[:-1].sum() // 255