`bincount` is faster than `histc` on CPU, so this should speed up (~2X) the `equalize` transforms.

This is sort of a follow-up to https://github.com/pytorch/vision/pull/3334 and https://github.com/pytorch/vision/issues/3173

On GPU:

```py
In [16]: img = torch.randint(0, 256, size=(460 * 680,)).to('cuda')

In [17]: %timeit torch.bincount(img, minlength=256)
1.52 ms ± 25.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [18]: %timeit torch.histc(img.to(torch.float32), bins=256, min=0, max=255)
127 µs ± 1.07 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

on CPU (different machine):

```py
img = torch.randint(0, 256, size=(460 * 680,))

%timeit torch.histc(img.to(torch.float32), bins=256, min=0, max=255)
696 µs ± 10.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

%timeit torch.bincount(img, minlength=256)
289 µs ± 1.06 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```