@fmassa 2021-03-03
There is a huge difference on CUDA. I think it would be worth opening an issue in PyTorch to point out that the implementation of `bincount` is suboptimal?

Also, it would be good to check the number of CUDA again, but this time using `torch.cuda.synchronize()` to get more accurate timings