@NicolasHug 2021-03-08
I created a new "Improvement" label, in accordance with https://github.com/pytorch/vision/issues/3351#issuecomment-791482582